4th Update
Coding the second encounter is proving to be more challenging than expected. Tackling three different enemies with different health values and making sure the story fits what's being done in the code without any errors or holes in logic is both time-consuming and straining. What coding this encounter has forced me to do, however, is to think about situations in a game through a logic of checks and numbers. Now when I'm looking at something like "X enemy dies" in my code, I see the values being changed in the code and not just the words being printed. I've completed the bulk of the work on this segment though, and the rest of the code from here on out should be much more straightforward now that I've tackled this challenge.