5th Update
This update has a complete version of the second enemy encounter. This took much longer to code than I had originally thought, and as detailed in my last update, I ran into multiple issues with coding this segment. It's now complete and (hopefully) bug free though, so I can move on to the final two segments of the text adventure, which I know for sure will be less difficult to code.