6th Update
The final update for the game, adding in the Chest puzzle and Ogre encounter, rounding out the entire game. As I predicted, coding these last two sections were not much of a hassle compared to the previous encounter I had coded, and I was able to complete these sections efficiently. Most of my code should contain appropriate comments for each section as well, including a hidden cheat to skip to the end (though I added this to bug test it, not for player use).